requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
requests
numpy
pandas
tensorflow
tensorflow-recommenders